+++ derive/tests/resolved_object/resolved_object_result_tests.rs
@@ -1,6 +1,7 @@
 use dynamic_graphql::dynamic::DynamicRequestExt;
 use dynamic_graphql::App;
 use dynamic_graphql::FieldValue;
+use dynamic_graphql::SimpleObject;
 use dynamic_graphql::{ResolvedObject, ResolvedObjectFields};
 
 use crate::schema_utils::normalize_schema;
@@ -98,6 +99,141 @@ async fn test_query() {
     );
 
     let root = Query { fail: true };
+    let req = dynamic_graphql::Request::new(query).root_value(FieldValue::owned_any(root));
+    let res = schema.execute(req).await;
+    assert_eq!(res.data.into_json().unwrap(), serde_json::json!(null));
+    assert_eq!(res.errors.len(), 1);
+    assert_eq!(res.errors[0].message, "Not found");
+}
+
+#[tokio::test]
+async fn test_ref_query() {
+    #[derive(ResolvedObject)]
+    #[graphql(root)]
+    struct Query {
+        fail: bool,
+        value: String,
+    }
+
+    #[ResolvedObjectFields]
+    impl Query {
+        async fn string(&self) -> Result<&String, MyError> {
+            if self.fail {
+                Err(MyError::NotFound)
+            } else {
+                Ok(&self.value)
+            }
+        }
+        async fn maybe_string(&self) -> Result<Option<&String>, MyError> {
+            if self.fail {
+                Err(MyError::NotFound)
+            } else {
+                Ok(Some(&self.value))
+            }
+        }
+    }
+
+    #[derive(App)]
+    struct App(Query);
+
+    let schema = App::create_schema().finish().unwrap();
+
+    let query = r#"
+        query {
+            string
+            maybeString
+        }
+    "#;
+    let root = Query {
+        fail: false,
+        value: "Hello".to_string(),
+    };
+    let req = dynamic_graphql::Request::new(query).root_value(FieldValue::owned_any(root));
+    let res = schema.execute(req).await;
+    assert_eq!(
+        res.data.into_json().unwrap(),
+        serde_json::json!({
+            "string": "Hello",
+            "maybeString": "Hello",
+        })
+    );
+
+    let root = Query {
+        fail: true,
+        value: "Hello".to_string(),
+    };
+    let req = dynamic_graphql::Request::new(query).root_value(FieldValue::owned_any(root));
+    let res = schema.execute(req).await;
+    assert_eq!(res.data.into_json().unwrap(), serde_json::json!(null));
+    assert_eq!(res.errors.len(), 1);
+    assert_eq!(res.errors[0].message, "Not found");
+}
+
+#[tokio::test]
+async fn test_ref_object_query() {
+    #[derive(SimpleObject)]
+    struct Foo {
+        value: String,
+    }
+
+    #[derive(ResolvedObject)]
+    #[graphql(root)]
+    struct Query {
+        fail: bool,
+        value: Foo,
+    }
+
+    #[ResolvedObjectFields]
+    impl Query {
+        async fn foo(&self) -> Result<&Foo, MyError> {
+            if self.fail {
+                Err(MyError::NotFound)
+            } else {
+                Ok(&self.value)
+            }
+        }
+        async fn maybe_foo(&self) -> Result<Option<&Foo>, MyError> {
+            if self.fail {
+                Err(MyError::NotFound)
+            } else {
+                Ok(Some(&self.value))
+            }
+        }
+    }
+
+    #[derive(App)]
+    struct App(Query);
+
+    let schema = App::create_schema().finish().unwrap();
+
+    let query = r#"
+        query {
+            foo { value }
+            maybeFoo { value }
+        }
+    "#;
+    let root = Query {
+        fail: false,
+        value: Foo {
+            value: "Hello".to_string(),
+        },
+    };
+    let req = dynamic_graphql::Request::new(query).root_value(FieldValue::owned_any(root));
+    let res = schema.execute(req).await;
+    assert_eq!(
+        res.data.into_json().unwrap(),
+        serde_json::json!({
+            "foo": { "value": "Hello" },
+            "maybeFoo": { "value": "Hello" },
+        })
+    );
+
+    let root = Query {
+        fail: true,
+        value: Foo {
+            value: "Hello".to_string(),
+        },
+    };
     let req = dynamic_graphql::Request::new(query).root_value(FieldValue::owned_any(root));
     let res = schema.execute(req).await;
     assert_eq!(res.data.into_json().unwrap(), serde_json::json!(null));